+++ client/src/app/(withauth)/products/pay/page.tsx
@@ -1,0 +1,110 @@
+"use client"
+import Hero from "@/components/hero-component"
+import Loading from "@/components/loading-component"
+import Show from "@/components/show-component"
+import { endPoints } from "@/constants/api-endpoints"
+import { uiConstants } from "@/constants/global-constants"
+import HTTPMethods from "@/constants/http-methods"
+import { GlobalContext } from "@/context/globalstate.provider"
+import useFetch from "@/hooks/use-fetch"
+import { ArrowRightIcon } from "@radix-ui/react-icons"
+import axios from "axios"
+import { useContext, useState } from "react"
+import { Badge, Button, Col, Container, Form, Row } from "react-bootstrap"
+import { toast } from "react-hot-toast"
+import Web3 from "web3"
+
+export default function Page() {
+  const products = useFetch("get-products", endPoints.getProductConfig, HTTPMethods.POST, { searchQuery: "pay" })
+  const selectedProduct = products?.data?.find((product: any) => product.productName === "pay")
+  const web3Provider = new Web3(endPoints.paySignTransactionGateway)
+  const [{ userState }] = useContext(GlobalContext)
+  const [matic, setMatic] = useState(0)
+  const [receiverAddress, setReceiverAddress] = useState("")
+  const [isLoading, setLoading] = useState(false)
+  const { address: walletAddress } = web3Provider.eth.accounts.privateKeyToAccount(userState.privateKey)
+
+  const sendMatic = async (e: any) => {
+    e.preventDefault()
+
+    try {
+      setLoading(true)
+      await axios.post(endPoints.payCreateTx)
+      const gasPrice = await web3Provider.eth.getGasPrice()
+
+      const transactionObject = {
+        from: walletAddress,
+        to: receiverAddress,
+        value: web3Provider.utils.toWei(matic.toString(), "ether"),
+        gas: 40000,
+        gasPrice: gasPrice,
+      }
+
+      const signedApprovalTx = await web3Provider.eth.accounts.signTransaction(transactionObject, userState.privateKey)
+
+      if (signedApprovalTx.rawTransaction) {
+        await web3Provider.eth.sendSignedTransaction(signedApprovalTx.rawTransaction)
+        toast.success(uiConstants.transactionSuccess)
+      }
+
+      else {
+        toast.error(uiConstants.transactionError)
+      }
+    }
+
+    catch (error: any) {
+      if (error.response && error.response.data.message) {
+        toast.error(error.response.data.message)
+      }
+
+      else {
+        toast.error(uiConstants.transactionError)
+      }
+    }
+
+    finally {
+      setLoading(false)
+    }
+  }
+
+  return (
+    <Container>
+      <Show when={!products.isLoading}>
+        <Hero>
+          <p className="branding">{uiConstants.brandName} {selectedProduct?.displayName}</p>
+          <p className="muted-text mt-3">{selectedProduct?.largeDescription}</p>
+          <div className="mb-2">
+            <Badge bg="light" className="mt-2 me-2 top-0 end-0 ps-3 pe-3 p-2">{selectedProduct?.productCategory}</Badge>
+            <Badge bg="light" className="mt-2 me-2 top-0 end-0 ps-3 pe-3 p-2">{selectedProduct?.productStatus}</Badge>
+          </div>
+          <form onSubmit={sendMatic}>
+            <p className="muted-text mt-4">Enter the details to send matic</p>
+            <Row className="g-2">
+              <Col xs={12} sm={12} md={6} lg={4} xl={3}>
+                <Form.Group className="mb-2" controlId="exampleForm.ControlInput1">
+                  <Form.Label>Wallet address</Form.Label>
+                  <Form.Control disabled={isLoading} autoFocus type="text" placeholder="Ethereum Wallet Address" onChange={(e) => setReceiverAddress(e.target.value)} required autoComplete={"off"} />
+                </Form.Group>
+              </Col>
+              <Col xs={12} sm={12} md={6} lg={4} xl={3}>
+                <Form.Group className="mb-3" controlId="exampleForm.ControlInput1">
+                  <Form.Label>MATIC Amount</Form.Label>
+                  <Form.Control disabled={isLoading} type="text" placeholder="MATIC Amount" onChange={(e) => setMatic(Number(e.target.value))} required autoComplete={"off"} />
+                </Form.Group>
+              </Col>
+            </Row>
+            <Col xs={12} sm={12} md={12} lg={8} xl={6}>
+              <Button variant="warning" type="submit" disabled={isLoading} className="mt-2 btn-block">
+                <Show when={!isLoading}>Send {matic} MATIC <ArrowRightIcon className="icon-right" /></Show>
+                <Show when={isLoading}><i className="fas fa-circle-notch fa-spin"></i> Sending MATIC</Show>
+              </Button>
+            </Col>
+          </form>
+        </Hero>
+      </Show>
+      <Show when={!!products.isLoading}>
+        <Loading />
+      </Show>
+    </Container>
+  )
+}