+++ client/src/app/loading.tsx
@@ -1,9 +1,14 @@
 "use client"
+import Header from "@/components/header-component"
+import { Fragment } from "react"
 
 export default function Loading() {
   return (
-    <div className="loading-container text-center">
-      <i className="fas fa-spinner fa-spin fa-3x"></i>
-    </div>
+    <Fragment>
+      <Header isAuthorized={false} />
+      <div className="loading-container text-center">
+        <i className="fas fa-spinner fa-spin fa-3x"></i>
+      </div>
+    </Fragment>
   )
 }